linkdin track assignmet/assignment-1/requirements.txt
fastapi
uvicorn
starlette
python-dotenv
authlib
celery
pydantic
pyjwt